+++ CHANGELOG.md
@@ -24,6 +24,7 @@ in the detailed section referring to by linking pull requests or issues.
 * Fix TODO and document `:extensions:data-plane-transfer` (#1519)
 * CloudEvents Http extension (#1160)
 * Full Query capabilities for SQL TransferProcess Store (#1483)
+* Full Query capabilities for SQL PolicyDefinition Store (#1484)
 
 #### Changed